+++ src/Command/ApiApplicationsCommand.php
@@ -3,10 +3,8 @@
 namespace App\Command;
 
 use App\Service\ApiClient;
-use Cerbero\JsonParser\JsonParser;
 use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
-use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
@@ -14,12 +12,13 @@ use Symfony\Component\Console\Style\SymfonyStyle;
 
 #[AsCommand(
     name: 'app:api:applications',
-    description: 'Add a short description for your command',
+    description: 'Get applications from API server',
 )]
 class ApiApplicationsCommand extends Command
 {
     public function __construct(
         private readonly ApiClient $apiClient,
+        private readonly array $applicationStatus,
     )
     {
         parent::__construct();
@@ -28,23 +27,29 @@ class ApiApplicationsCommand extends Command
     protected function configure(): void
     {
         $this
-            ->addArgument('arg1', InputArgument::OPTIONAL, 'Argument description')
-            ->addOption('option1', null, InputOption::VALUE_NONE, 'Option description')
+            ->addOption('filterStatus', null, InputOption::VALUE_NONE, 'Filter based on configured statuses: ' . implode(',', $this->applicationStatus))
         ;
     }
 
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
         $io = new SymfonyStyle($input, $output);
-        $arg1 = $input->getArgument('arg1');
+        $filter = $input->getOption('filterStatus');
 
-        $apps = $this->apiClient->getApplications();
+        $apps = $this->apiClient->getApplications($filter);
 
         foreach ($apps as $app) {
             $io->writeln('<info>'.$app->name.'</info>');
         }
 
-        $io->success(count($apps) . ' applications found');
+        $msg = count($apps);
+        if ($filter) {
+            $msg .= sprintf(' applications found (filter on status "%s")', implode(',', $this->applicationStatus));
+        }
+        else {
+            $msg .= ' applications found';
+        }
+        $io->success($msg);
 
         return Command::SUCCESS;
     }